requirements.txt
absl-py==0.9.0
astor==0.8.1
Click==7.0
Flask==1.1.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.27.2
gunicorn==20.0.4
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.2.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
opt-einsum==3.2.0
Pillow==7.0.0
protobuf==3.11.3
PyYAML==5.3
scipy==1.4.1
six==1.14.0
tensorboard==1.15.0
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
termcolor==1.1.0
Werkzeug==1.0.0
wrapt==1.12.0
